# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação API controle de acesso Laravel",
        "description": "Documentação com todos os endpoints necessários para a correta utilização da aplicação.",
        "contact": {
            "email": "assuncaofilipe97@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "",
            "description": "Local"
        }
    ],
    "paths": {
        "/app/login": {
            "post": {
                "tags": [
                    "Acesso"
                ],
                "summary": "Autenticar usuário",
                "description": "Retorna usuário logado",
                "operationId": "342ade8140987f5ceae4440337c216fe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            },
                            "examples": {
                                "register": {
                                    "summary": "Login",
                                    "value": {
                                        "email": "exemplo@gmail.com",
                                        "password": "123a123a"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login efetuado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/loginResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Senha incorreta, favor revisar",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": "false",
                                    "data": {
                                        "message": "Senha incorreta, favor revisar."
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Campos inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "email": {
                                                    "type": "string"
                                                },
                                                "password": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "examples": {
                                    "login": {
                                        "summary": "Campos obrigatórios",
                                        "value": {
                                            "success": "false",
                                            "data": {
                                                "email": [
                                                    "O campo email é obrigatório."
                                                ],
                                                "password": [
                                                    "O campo senha é obrigatório."
                                                ]
                                            }
                                        }
                                    },
                                    "login2": {
                                        "summary": "Campos inválidos",
                                        "value": {
                                            "success": "false",
                                            "data": {
                                                "email": [
                                                    "Email não cadastrado"
                                                ],
                                                "password": [
                                                    "O campo senha deve ter pelo menos 8 caracteres"
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/app/logout": {
            "post": {
                "tags": [
                    "Acesso"
                ],
                "summary": "Desconectar usuário",
                "description": "Desconecta um usuário conectado",
                "operationId": "45468c7cb12267053791dbe5825d037e",
                "responses": {
                    "200": {
                        "description": "Usuário Desconectado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": "true",
                                    "data": {
                                        "message": "Usuário desconectado com sucesso"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/forbidden"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/app/refresh": {
            "post": {
                "tags": [
                    "Acesso"
                ],
                "summary": "Obter novo token",
                "description": "Retorna informações do usuário e um novo token",
                "operationId": "6d6cf921991439f285f6f40acf213c9d",
                "responses": {
                    "200": {
                        "description": "Token atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/loginResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/forbidden"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/app/me": {
            "get": {
                "tags": [
                    "Acesso"
                ],
                "summary": "Obter dados do usuário conectado",
                "description": "Retorna informações do usuário conectado",
                "operationId": "26cc16fd7bf378902db25ce59c26e740",
                "responses": {
                    "200": {
                        "description": "Dados do usuário conectado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "": {
                                                    "$ref": "#/components/schemas/user"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": "true",
                                    "data": {
                                        "email": "teste@gmail.comffffffffff",
                                        "uuid": "fbd5d732-137b-462b-bfd1-d32b23209fa5",
                                        "updated_at": "2022-02-09T19:32:49.000000Z",
                                        "created_at": "2022-02-09T19:32:49.000000Z",
                                        "id": 470
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/forbidden"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "loginResponse": {
                "properties": {
                    "success": {
                        "type": "string"
                    },
                    "data": {
                        "properties": {
                            "access_token": {
                                "type": "string"
                            },
                            "token_type": {
                                "type": "string"
                            },
                            "expires_in": {
                                "type": "string",
                                "format": "date"
                            },
                            "user": {
                                "$ref": "#/components/schemas/user"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "example": {
                    "success": "true",
                    "data": {
                        "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC90cm91dy1hcHBcL2FwcFwvYXV0aFwvbG9naW4iLCJpYXQiOjE2NDQzMjYwODksImV4cCI6MTY0NDMyOTY4OSwibmJmIjoxNjQ0MzI2MDg5LCJqdGkiOiJDeENTS0x6cGRLQ3ZUc2ZQIiwic3ViIjoxOTUsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.D8i4qR2vgW9G-lz-ej1swDUToFgU4AnKHEFY_2hWRGA",
                        "token_type": "bearer",
                        "expires_in": 3600,
                        "user": {
                            "id": 195,
                            "uuid": "d83810de-4d4a-4306-85ed-c2e761fc998a",
                            "email": "teste2@gmail.com",
                            "email_verified_at": null,
                            "created_at": "2022-02-08T13:09:57.000000Z",
                            "updated_at": "2022-02-08T13:09:57.000000Z"
                        }
                    }
                }
            },
...
            "user": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "uuid": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "email_verified_at": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "forbidden": {
                "description": "Token de autorização não encontrado / Token expirado",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "string"
                                },
                                "data": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        },
                        "examples": {
                            "Token de autorização não encontrado": {
                                "summary": "Token de autorização não encontrado",
                                "value": {
                                    "success": "false",
                                    "data": {
                                        "message": "Token de autorização não encontrado"
                                    }
                                }
                            },
                            "Token expirado": {
                                "summary": "Token expirado",
                                "value": {
                                    "success": "false",
                                    "data": {
                                        "message": "Token expirado"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "unauthorized": {
                "description": "Token inválido",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "string"
                                },
                                "data": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        },
                        "example": {
                            "success": "false",
                            "data": {
                                "message": "Token inválido"
                            }
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}